Separate old Hint code to another file
